Fix for odoo pos module
By default, negative amounts cause the pos to use the refund side of tax settings ( tax case, tax account )

Neolabo does not want this because these negative entries don't occur on a refund sale journal, but in the regular sale journal for the pos.
This patch changes that behaviour, so that negative amounts are treated just the same as postitive amounts

Because a fix in a custom module requires to copy the entire 250 line "_create_account_move_line" method of the pos_order object i chose to use this patch on the standard odoo core.